+++ src/param.c
@@ -541,8 +541,11 @@ set_init(void)
 	curbuf->b_syn_matchend	= NULL;
 	curbuf->b_syn_curp		= NULL;
 	curbuf->b_syn_line		= -1;
-	curbuf->b_syn_nline		= 0;
 	curbuf->b_syn_link		= NULL;
+	curbuf->b_syn_state		= NULL;
+	curbuf->b_syn_statelen	= 0;
+	curbuf->b_syn_stateval	= 0;
+	curbuf->b_syn_pairs		= 0;
 #endif
 #ifdef NT
 	{